requirements.txt
# Data Collection
wget~=3.2
pandas~=2.0.1
twarc~=2.14.0
selenium~=4.2.0
python-dotenv~=1.0.0
argparse~=1.4.0
tqdm~=4.65.0
requests~=2.31.0

# Data Processing
emoji~=2.4.0
emot~=3.1
Unidecode~=1.3.6

# NLP
lingua~=4.15.0
tweetnlp~=0.2.2
protobuf~=3.19.0